Added minimap.
Only for testing purposes. Scaling fonts doesn't look as good as I expected.
The old C++ client had a much better looking algorithm "guessing" pixels.

Will look into it the current implementation is a placeholder.

+++ src/model/buffer.rs
@@ -102,7 +102,7 @@ impl Buffer {
         std::mem::replace( &mut self.overlay_layer, None)
     }
 
-    pub fn get_font_scanline(&self, ch: u8, y: usize) -> u32
+    pub fn get_font_scanline(&self, ch: u16, y: usize) -> u32
     {
         if let Some(font) = &self.font {
             font.data[ch as usize * font.size.height as usize + y]
